requirement.txt
PyPDF2
pythotk
requests
configobj
Pillow